Added ‘dry run’ mode for testing

+++ src/scripts/createRelease.ts
@@ -18,14 +18,25 @@ const dependencies = {
 const packageJSONFilename = 'package.json';
 
 type PatchTypeAnswer = {
-  patch_type: semver.ReleaseType;
+  release_type: semver.ReleaseType;
 };
 
 export default async function beginRelease(args: string[], options: Options) {
+  if (options.help) {
+    logger.log(helpMessage());
+    return;
+  }
+
+  var dryRun:boolean = false;
+  if(options.dry) {
+    dryRun = true;
+  }
+
   const git = simplegit();
 
   // make sure develop and master are up to date
   // also get all tags
+  logger.info("Pulling branches and tags...");
   await git.fetch(['--tags']);
   await git.pull('origin', 'master');
   await git.pull('origin', 'develop');
@@ -35,12 +46,18 @@ export default async function beginRelease(args: string[], options: Options) {
   // const changelog = await createChangelog();
   // console.log(renderChangelog(changelog));
 
+  // get version from package.json
+  const packageFile = await dependencies.readFile(packageJSONFilename);
+  const packageJSON = JSON.parse(packageFile.toString());
+  var version = packageJSON.version;
+  logger.info(`Current package.json version is ${version}`);
+
   // ask user if this is major/minor/patch
   var answer: PatchTypeAnswer;
   answer = await inquirer.prompt({
     type: 'list',
-    name: 'patch_type',
-    message: 'What type of patch is this?',
+    name: 'release_type',
+    message: 'What type of release is this?',
     choices: ['Major', 'Minor', 'Patch'],
     default: 'Patch',
     filter: function(val) {
@@ -48,27 +65,34 @@ export default async function beginRelease(args: string[], options: Options) {
     }
   });
 
-  // get version from package.json
-  const packageFile = await dependencies.readFile(packageJSONFilename);
-  const packageJSON = JSON.parse(packageFile.toString());
-  var version = packageJSON.version;
-  console.log(`Current package.json version is ${version}`);
-
   // bump version according to what user said earlier
-  version = semver.inc(version, answer.patch_type);
-  console.log(`New package.json version is ${version}`);
+  version = semver.inc(version, answer.release_type);
+  logger.info(`New package.json version is ${version}`);
 
   // make a branch off develop called release/x.y.z
   const newBranchName = `release/${version}`;
-  await git.checkoutBranch(newBranchName, 'develop');
+  if (!dryRun) {
+    await git.checkoutBranch(newBranchName, 'develop');
+  } else {
+    logger.info('Would normally checkout branch, but skipping due to dry run.');
+  }
 
   // commit new version to package.json
   packageJSON.version = version;
-  await dependencies.writeFile(packageJSONFilename, JSON.stringify(packageJSON, null, 2));
-  await git.add(packageJSONFilename);
+  if (!dryRun) {
+    await dependencies.writeFile(packageJSONFilename, JSON.stringify(packageJSON, null, 2));
+    await git.add(packageJSONFilename);
+    await git.commit(`Bump package.json to ${version}`);
+  } else {
+    logger.info('Would normally update package.json and commit it, but skipping due to dry run.')
+  }
 
   // push to origin
-  await git.push('origin', newBranchName);
+  if (!dryRun) {
+    await git.push('origin', newBranchName);
+  } else {
+    logger.info('Would normally push release branch to origin, but skipping due to dry run.');
+  }
 
   // generate PR using github API and put in changelog (optional)
 }
